# core/src/main/kotlin/com/iota/core/model/workflows/Workflow.kt
package com.iota.core.model.workflows

import jakarta.persistence.Entity
import jakarta.persistence.GeneratedValue
import jakarta.persistence.Id
import jakarta.persistence.ManyToMany
import jakarta.validation.constraints.NotEmpty

@Entity
class Workflow {
    @Id
    @GeneratedValue
    var id: Long = 0

    @NotEmpty
    var name: String = ""

    var active: Boolean = true

    @ManyToMany
    var entryNodes: List<EventNode> = listOf()
}